fix: treat time as unsigned 64 bit number

+++ log_parser/DFParser.py
@@ -231,7 +231,7 @@ class DFLog(object):
                     continue
                 temp_array = self.tables[table].to_numpy()
                 temp_array = np.hstack([temp_array[:, :1], 
-                                        np.reshape([int(float(val)) for val in temp_array[:, 1]], 
+                                        np.reshape([np.uint64(val) for val in temp_array[:, 1]], 
                                                    (len(temp_array), 1)),
                                         np.reshape([", ".join([str(x) 
                                                          for x in row]) 
@@ -241,7 +241,7 @@ class DFLog(object):
                     numpy_msgs = temp_array
                 else:
                     numpy_msgs = np.vstack((numpy_msgs, temp_array))
-            numpy_msgs = numpy_msgs[numpy_msgs[:, 1].astype(float).argsort()]
+            numpy_msgs = numpy_msgs[numpy_msgs[:, 1].astype(np.uint64).argsort()]
             np.savetxt(outfile, numpy_msgs, fmt='%s', delimiter=', ', newline='\n')
 
     def drop_empty(self):
@@ -309,7 +309,7 @@ class DFLog(object):
         
         # and insert the new message dataframes into tables
         for name in [x for x in other.tables if x not in drop_tables and x not in format_table_names]:
-            other.tables[name]['TimeUS'] = other.tables[name]['TimeUS'].astype(int) + time_shift
+            other.tables[name]['TimeUS'] = other.tables[name]['TimeUS'].astype(np.uint64) + time_shift
             self.tables[name] = other.tables[name]
     
     def find_offset(self, other):
